requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
lxml==4.6.2
    # via python-docx
python-docx==0.8.10
    # via -r requirements.in
